requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
soupsieve>=2.5
transformers>=4.41.0
fastembed>=0.3.1
faiss-cpu>=1.7.4
numpy>=1.26.0
flask>=3.0.0

